refac: Refactor InsurancePlan classes
Creates a method to return the RiskProfileHandler chain of
responsibility.
This way, it is not necessary to keep the RiskRuleHandlerFactory class,
that has some complex and hard to understand code. This way might be
more coupled, but I find it easier recognize the code purpose and it
makes it easier to test.